requirements.txt
click
docker
pyyaml